:recycle: #52 exceptionsをExceptionsディレクトリに名前を変更
