+++ app/Models/Domain/Exceptions/BusinessLogicException.php
@@ -3,7 +3,7 @@
  * BusinessLogicException
  */
 
-namespace App\Models\Domain\exceptions;
+namespace App\Models\Domain\Exceptions;
 
 /**
  * Class BusinessLogicException
