+++ app/Models/Domain/Exceptions/LoginSessionExpiredException.php
@@ -3,7 +3,7 @@
  * LoginSessionExpiredException
  */
 
-namespace App\Models\Domain\exceptions;
+namespace App\Models\Domain\Exceptions;
 
 use Throwable;
