+++ app/Services/AccountScenario.php
@@ -14,11 +14,11 @@ use App\Models\Domain\AccountSpecification;
 use App\Models\Domain\LoginSessionRepository;
 use App\Models\Domain\QiitaAccountValueBuilder;
 use App\Models\Domain\LoginSessionEntityBuilder;
-use App\Models\Domain\exceptions\ValidationException;
-use App\Models\Domain\exceptions\UnauthorizedException;
+use App\Models\Domain\Exceptions\ValidationException;
+use App\Models\Domain\Exceptions\UnauthorizedException;
 use Illuminate\Database\Eloquent\ModelNotFoundException;
-use App\Models\Domain\exceptions\AccountCreatedException;
-use App\Models\Domain\exceptions\LoginSessionExpiredException;
+use App\Models\Domain\Exceptions\AccountCreatedException;
+use App\Models\Domain\Exceptions\LoginSessionExpiredException;
 
 /**
  * Class AccountScenario
